+++ examples/gh-pages/src/vite-env.d.ts
@@ -1,5 +1,7 @@
 /// <reference types="vite/client" />
 
+declare const __REPL_VERSION__: string;
+
 declare module '*.wasm?url' {
   const url: string;
   export default url;